+++ plugins/nf-quilt/src/test/nextflow/quilt/jep/QuiltPackageTest.groovy
@@ -22,6 +22,7 @@ import nextflow.quilt.nio.QuiltPath
 
 import spock.lang.Ignore
 import spock.lang.IgnoreIf
+import spock.lang.Unroll
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
@@ -63,6 +64,25 @@ class QuiltPackageTest extends QuiltSpecification {
         pkg == pkg2
     }
 
+    @Unroll
+    void 'should find relativeChildren of complete folders'() {
+        given:
+        Path path = Paths.get(new URI(PACKAGE_URL))
+        when:
+        QuiltPackage pkg = path.pkg()
+        then:
+        pkg.relativeChildren(subpath).size() == expected_size
+        
+        where:
+        subpath               | expected_size
+        ''                    | 8
+        '.ipynb_checkpoints'  | 1
+        '.ipynb_checkpoints/' | 1
+        '.ipynb'              | 1
+        '.ipynb/'             | 0
+
+    }
+
     void 'should distinguish Packages with same name in different Buckets '() {
         given:
         def url2 = TEST_URL.replace('quilt-', 'quilted-')
@@ -86,6 +106,30 @@ class QuiltPackageTest extends QuiltSpecification {
         Files.exists(installPath)
     }
 
+    void 'should copy temp files into install folder'() {
+        given:
+        String filename = 'test.txt'
+        Path installPath = pkg.packageDest()
+        Path tempFile = File.createTempFile('test', '.txt').toPath()
+        Path installedFile = Paths.get(installPath.toString(), filename)
+        expect:
+        Files.exists(tempFile)
+        Files.exists(installPath)
+        !Files.exists(installedFile)
+        Files.copy(tempFile, installedFile)
+        Files.exists(installedFile)
+    }
+
+    void 'should copy package files to temp Path'() {
+        given:
+        Path installPath = pkg.packageDest()
+        expect:
+        Files.exists(installPath)
+        Files.isDirectory(installPath)
+        Files.readAttributes(installPath, BasicFileAttributes)
+    }
+
+
     void 'should get attributes for package folder'() {
         given:
         def root = qpath.getRoot()
@@ -106,6 +150,7 @@ class QuiltPackageTest extends QuiltSpecification {
         Files.readAttributes(qpath, BasicFileAttributes)
     }
 
+
     void 'should return null on failed install'() {
         given:
         def url2 = TEST_URL.replace('quilt-', 'quilted-')
@@ -119,18 +164,14 @@ class QuiltPackageTest extends QuiltSpecification {
     @IgnoreIf({ System.getProperty('os.name').contains('indows') })
     void 'should deinstall files'() {
         expect:
-        Files.exists(qpath.localPath())
+        Files.exists(qpath.localPath(true))
+        Files.readAttributes(qpath, BasicFileAttributes)
         when:
         qpath.deinstall()
         then:
-        !Files.exists(qpath.localPath())
-        /* when:
-        Files.readAttributes(qpath, BasicFileAttributes)
-        then:
-        thrown(java.nio.file.NoSuchFileException) */
+        !Files.exists(qpath.localPath(false))
     }
 
-    @Ignore()
     void 'should iterate over installed files '() {
         given:
         def root = qpath.getRoot()
@@ -192,6 +233,7 @@ class QuiltPackageTest extends QuiltSpecification {
         opkg.push('msg', meta) == 0
     }
 
+    @Ignore('QuiltCore-java does not support workflows yet')
     @IgnoreIf({ env.WRITE_BUCKET == 'quilt-example' || env.WRITE_BUCKET ==  null })
     void 'should fail if invalid workflow'() {
         given:
@@ -201,6 +243,7 @@ class QuiltPackageTest extends QuiltSpecification {
         bad_wf.push('missing-workflow first time', [:]) == 1
     }
 
+    @Ignore('QuiltCore-java does not support workflows yet')
     @IgnoreIf({ env.WRITE_BUCKET == 'quilt-example' || env.WRITE_BUCKET ==  null })
     void 'should fail push if unsatisfied workflow'() {
         given: