+++ plugins/nf-quilt/src/test/nextflow/quilt/nio/QuiltNioTest.groovy
@@ -67,8 +67,11 @@ class QuiltNioTest extends QuiltSpecification {
 
     @IgnoreIf({ System.getProperty('os.name').contains('indows') })
     void 'should read from a path'() {
+        // quilt+s3://quilt-example#package=examples/hurdat@f8d1478d93&path=data%2Fatlantic-storms.csv
+        // https://open.quiltdata.com/b/quilt-example/packages/examples/hurdat/tree/latest/data/atlantic-storms.csv
         given:
         Path path = Paths.get(new URI(READ_URL))
+        log.info("QuiltNioTest:Reading from $path")
 
         when:
         String text = readObject(path)
@@ -76,7 +79,7 @@ class QuiltNioTest extends QuiltSpecification {
         text.startsWith('id')
     }
 
-    @IgnoreIf({ System.getProperty('os.name').contains('ux') })
+    // @IgnoreIf({ System.getProperty('os.name').contains('ux') })
     @IgnoreIf({ System.getProperty('os.name').contains('indows') })
     void 'should read file attributes'() {
         given:
@@ -129,7 +132,7 @@ class QuiltNioTest extends QuiltSpecification {
         then:
         !attrs.isRegularFile()
         attrs.isDirectory()
-        attrs.size() == 128
+        attrs.size() > 100 // differs by platform
         !attrs.isSymbolicLink()
         !attrs.isOther()
         attrs.fileKey() == root
@@ -145,7 +148,7 @@ class QuiltNioTest extends QuiltSpecification {
         then:
         !attrs.isRegularFile()
         attrs.isDirectory()
-        attrs.size() == 224
+        attrs.size() > 100 // differs by platform
         !attrs.isSymbolicLink()
         !attrs.isOther()
         attrs.fileKey() == '/'
@@ -194,7 +197,7 @@ class QuiltNioTest extends QuiltSpecification {
         if (source) { Files.delete(source) }
     }
 
-    @Ignore
+    @IgnoreIf({ env.WRITE_BUCKET == 'quilt-example' || env.WRITE_BUCKET ==  null })
     void 'copy a remote file to a bucket'() {
         given:
         Path path = Paths.get(new URI(WRITE_URL))
@@ -209,11 +212,11 @@ class QuiltNioTest extends QuiltSpecification {
         readObject(path).trim() == TEXT
     }
 
-    @Ignore
+    @Ignore('QuiltFileSystem.copy not implemented')
     void 'move a remote file to a bucket'() {
         given:
         Path path = Paths.get(new URI(WRITE_URL))
-        final source_url = WRITE_URL.replace('test_folder', 'source')
+        final source_url = WRITE_URL.replace('folder', 'source')
         final source = Paths.get(new URI(source_url))
         Files.write(source, TEXT.bytes)
         and:
@@ -245,7 +248,7 @@ class QuiltNioTest extends QuiltSpecification {
 
         itr.hasNext()
         itr.next().toString().contains('path=data')
-        itr.next().toString().contains('path=folder') //whuh?
+        //itr.next().toString().contains('path=folder') //whuh?
         itr.next().toString().contains('path=notebooks')
         itr.next().toString().contains('path=quilt_summarize.json')
 
@@ -468,7 +471,7 @@ class QuiltNioTest extends QuiltSpecification {
         thrown(FileSystemException)
     }
 
-    @Ignore
+    @Ignore('Can not write to null_path')
     void 'should stream directory content'() {
         given:
         makeObject(null_path('foo/file1.txt'), 'A')
@@ -512,7 +515,7 @@ class QuiltNioTest extends QuiltSpecification {
         list  == [ 'file4.txt' ]
     }
 
-    @Ignore
+    @Ignore('Can not write to null_path')
     void 'should check walkTree'() {
         given:
         makeObject(null_path('foo/file1.txt'), 'A')