+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.5.0] 2023-09-04+
+
+- Switch to quiltcore-java [NOTE: this pre-release does NOT check workflows]
+- Do not pre-install packages (only install before write)
+
 ## [0.4.5] 2023-08-23
 
 - fix metadata in README.md